Working socket acceptor
Had to add IDs to the socket acceptor child_spec as they need to be unique.
SocketAcceptor no longer errors and exists after getting a connection it
doesn't like. Before this the SocketAcceptor would die and require a restart,
and if the connections came fast enough the SocketSupervisor would die too.
I'm unsure why that was happening, maybe the listening socket was going out
of scope when all the child processes were dead. In any case we use a `with`
clause now to suppress the errors and keep the acceptor from exiting.